fix(line2) renamed intersectToLine -> intersectPointOfLines
