+++ src/core/math/line2/intersectPointOfLines.js
